修复删除视频时无论取消还是确认都会删除视频的BUG

+++ src/views/Course.vue
@@ -181,8 +181,8 @@
         <el-table-column prop="school_id" label="学号"></el-table-column>
         <el-table-column prop="realname" label="姓名"></el-table-column>
         <el-table-column prop="school" label="学校"></el-table-column>
-        <el-table-column  label="总观看次数"></el-table-column>
-        <el-table-column  label="总播放时长"></el-table-column>
+        <el-table-column label="总观看次数"></el-table-column>
+        <el-table-column label="总播放时长"></el-table-column>
         <el-table-column prop="user_identity" label="身份">
           <template slot-scope="scope">
             <span v-if="scope.row.user_identity == 0">学生</span>
@@ -573,9 +573,11 @@ export default {
       this.$alert("确定删除视频吗？", "删除视频", {
         confirmButtonText: "确定",
         callback: async (action) => {
-          const temp = await CourseAPI.deleteVideo(video.id);
-          console.log("删除返回");
-          console.log(temp);
+          if (action === "confirm") {
+            const temp = await CourseAPI.deleteVideo(video.id);
+            console.log("删除返回");
+            console.log(temp);
+          }
         },
       });
     },
